% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucells.R
\name{AUCell_calcAUC}
\alias{AUCell_calcAUC}
\alias{AUCell_calcAUC,list-method}
\alias{AUCell_calcAUC,character-method}
\alias{AUCell_calcAUC,GeneSet-method}
\alias{AUCell_calcAUC,GeneSetCollection-method}
\title{Calculate AUC}
\usage{
AUCell_calcAUC(
  geneSets,
  rankings,
  nCores = 1,
  normAUC = TRUE,
  aucMaxRank = ceiling(0.05 * nrow(rankings)),
  verbose = TRUE
)

\S4method{AUCell_calcAUC}{list}(
  geneSets,
  rankings,
  nCores = 1,
  normAUC = TRUE,
  aucMaxRank = ceiling(0.05 * nrow(rankings)),
  verbose = TRUE
)

\S4method{AUCell_calcAUC}{character}(
  geneSets,
  rankings,
  nCores = 1,
  normAUC = TRUE,
  aucMaxRank = ceiling(0.05 * nrow(rankings)),
  verbose = TRUE
)

\S4method{AUCell_calcAUC}{GeneSet}(
  geneSets,
  rankings,
  nCores = 1,
  normAUC = TRUE,
  aucMaxRank = ceiling(0.05 * nrow(rankings)),
  verbose = TRUE
)

\S4method{AUCell_calcAUC}{GeneSetCollection}(
  geneSets,
  rankings,
  nCores = 1,
  normAUC = TRUE,
  aucMaxRank = ceiling(0.05 * nrow(rankings)),
  verbose = TRUE
)
}
\arguments{
\item{geneSets}{List of gene-sets (or signatures) to test in the cells.
The gene-sets should be provided as \code{\link{GeneSet}},
\code{\link{GeneSetCollection}} or character list (see examples).}

\item{rankings}{'Rankings' created for this dataset with
\code{\link{AUCell_buildRankings}}.}

\item{nCores}{Number of cores to use for computation.}

\item{normAUC}{Wether to normalize the maximum possible AUC to 1 (Default: TRUE).}

\item{aucMaxRank}{Threshold to calculate the AUC (see 'details' section)}

\item{verbose}{Should the function show progress messages? (TRUE / FALSE)}
}
\value{
Matrix with the AUC values (gene-sets as rows, cells as columns).
}
\description{
Calculates the 'AUC' for each gene-set in each cell.
}
\details{
In a simplified way, the AUC value represents the fraction of genes,
 within the top X genes in the ranking, that are included in the signature.
The parameter 'aucMaxRank' allows to modify the number of genes
(maximum ranking) that is used to perform this computation.
By default, it is set to 5\% of the total number of genes in the rankings.
Common values may range from 1 to 20\%.
}
\seealso{
Previous step in the workflow: \code{\link{AUCell_buildRankings}}.
Next step in the workflow: \code{\link{AUCell_exploreThresholds}}.

See the package vignette for examples and more details:
\code{vignette("AUCell")}
}
